refactor(core): split connections manager actor in different files
