+++ core/src/actors/connections_manager/messages.rs
@@ -1,0 +1,28 @@
+use std::net::SocketAddr;
+
+use actix::{actors::resolver::ResolverError, Message};
+use tokio::net::TcpStream;
+
+/// Actor message that holds the TCP stream from an inbound TCP connection
+#[derive(Message)]
+pub struct InboundTcpConnect {
+    /// Tcp stream of the inbound connections
+    pub stream: TcpStream,
+}
+
+impl InboundTcpConnect {
+    /// Method to create a new InboundTcpConnect message from a TCP stream
+    pub fn new(stream: TcpStream) -> InboundTcpConnect {
+        InboundTcpConnect { stream }
+    }
+}
+
+/// Actor message to request the creation of an outbound TCP connection to a peer.
+#[derive(Message)]
+pub struct OutboundTcpConnect {
+    /// Address of the outbound connection
+    pub address: SocketAddr,
+}
+
+/// Returned type by the Resolver actor for the ConnectAddr message
+pub type ResolverResult = Result<TcpStream, ResolverError>;
