+++ core/src/actors/sessions_manager.rs
@@ -14,7 +14,7 @@ use tokio::{codec::FramedRead, io::AsyncRead, net::TcpStream};
 
 use crate::actors::codec::P2PCodec;
 use crate::actors::config_manager::send_get_config_request;
-use crate::actors::connections_manager::{ConnectionsManager, OutboundTcpConnect};
+use crate::actors::connections_manager::{messages::OutboundTcpConnect, ConnectionsManager};
 use crate::actors::peers_manager::{AddPeers, GetRandomPeer, PeersManager, PeersSocketAddrResult};
 use crate::actors::session::{GetPeers, Session};
